+++ changelog.md
@@ -18,6 +18,8 @@ ex: JF: added robot images to the assets folder
 
 
 ### Changed
+
+- MG: Updated css to improve parts view, updated model in prep for next issue.
 - AR: update history model to include all types in every sub array
 - MG: Updated template to provide some code cleanup.
 - JF: Updated logo